E2Utility 1.5
Version 1.5 is Deployed. Please, redownload the dll.

​​​​​1. Sivir E is added on Ability Timer

1047

@sixi is requested

2. Fixed Possible issue not using QSS on Warwick R(Hopefully)

Tested video.

[SPOILER]

https://streamable.com/sn0zi

[/SPOILER]

1045

3. Cooldown Tracker is Added

[SPOILER]

1046

1048

[/SPOILER]

I had thought to add HUD on Cooldown tracker, but after I coded some basic functions of the Cooldown tracker, I don't think HUD actually needed.

You can customize positions of drawings, colors, time formats and so on.

The tracker itself is using Cooldown Expires to determine whether is the spell is on cooldown or not so it works most of spells.

However, while I was working on it, I found that some of spells were not tracked properly such as Lux's E, Rengar W(when no one nearby), Rumble's E. It is Riot's fault not me.

A good thing is I have implemented a function to be able to track the spells properly, but I have to manually insert the spell information.

To get to the point, if you see some of spells are not tracked properly, please let me know.

I will fix it.